--- packages/server/events-processing-platform/repository/organization_repository.go
@@ -1,316 +1,0 @@
-package repository
-
-import (
-	"context"
-	"fmt"
-	"github.com/neo4j/neo4j-go-driver/v5/neo4j"
-	"github.com/openline-ai/openline-customer-os/packages/server/customer-os-common-module/utils"
-	"github.com/openline-ai/openline-customer-os/packages/server/events-processing-platform/constants"
-	"github.com/openline-ai/openline-customer-os/packages/server/events-processing-platform/tracing"
-	"github.com/opentracing/opentracing-go"
-	"github.com/opentracing/opentracing-go/log"
-	"strings"
-	"time"
-)
-
-type OrganizationRepository interface {
-	LinkWithDomain(ctx context.Context, tenant, organizationId, domain string) error
-	ReplaceOwner(ctx context.Context, tenant, organizationId, userId string) error
-	SetVisibility(ctx context.Context, tenant, organizationId string, hide bool) error
-	UpdateLastTouchpoint(ctx context.Context, tenant, organizationId string, touchpointAt time.Time, touchpointId, touchpointType string) error
-	SetCustomerOsIdIfMissing(ctx context.Context, tenant, organizationId, customerOsId string) error
-	LinkWithParentOrganization(ctx context.Context, tenant, organizationId, parentOrganizationId, subOrganizationType string) error
-	UnlinkParentOrganization(ctx context.Context, tenant, organizationId, parentOrganizationId string) error
-	UpdateArr(ctx context.Context, tenant, organizationId string) error
-	UpdateRenewalSummary(ctx context.Context, tenant, organizationId string, likelihood *string, likelihoodOrder *int64, nextRenewalDate *time.Time) error
-	WebScrapeRequested(ctx context.Context, tenant, organizationId, url string, attempt int64, requestedAt time.Time) error
-	UpdateOnboardingStatus(ctx context.Context, tenant, organizationId, status, comments string, statusOrder *int64, updatedAt time.Time) error
-}
-
-type organizationRepository struct {
-	driver   *neo4j.DriverWithContext
-	database string
-}
-
-func NewOrganizationRepository(driver *neo4j.DriverWithContext, database string) OrganizationRepository {
-	return &organizationRepository{
-		driver:   driver,
-		database: database,
-	}
-}
-
-func (r *organizationRepository) LinkWithDomain(ctx context.Context, tenant, organizationId, domain string) error {
-	span, ctx := opentracing.StartSpanFromContext(ctx, "OrganizationRepository.MergeOrganizationDomain")
-	defer span.Finish()
-	tracing.SetNeo4jRepositorySpanTags(ctx, span, tenant)
-	span.LogFields(log.String("organizationId", organizationId))
-
-	cypher := `MERGE (d:Domain {domain:$domain}) 
-				ON CREATE SET 	d.id=randomUUID(), 
-								d.createdAt=$now, 
-								d.updatedAt=$now,
-								d.appSource=$appSource
-				WITH d
-				MATCH (t:Tenant {name:$tenant})<-[:ORGANIZATION_BELONGS_TO_TENANT]-(org:Organization {id:$organizationId})
-				MERGE (org)-[rel:HAS_DOMAIN]->(d)
-				SET rel.syncedWithEventStore = true
-				RETURN rel`
-	params := map[string]any{
-		"tenant":         tenant,
-		"organizationId": organizationId,
-		"domain":         strings.ToLower(domain),
-		"appSource":      constants.AppSourceEventProcessingPlatform,
-		"now":            utils.Now(),
-	}
-	span.LogFields(log.String("cypher", cypher))
-	tracing.LogObjectAsJson(span, "params", params)
-
-	session := utils.NewNeo4jWriteSession(ctx, *r.driver, utils.WithDatabaseName(r.database))
-	defer session.Close(ctx)
-
-	_, err := session.ExecuteWrite(ctx, func(tx neo4j.ManagedTransaction) (any, error) {
-		queryResult, err := tx.Run(ctx, cypher, params)
-		if err != nil {
-			return nil, err
-		}
-		_, err = queryResult.Single(ctx)
-		if err != nil {
-			return nil, err
-		}
-		return nil, nil
-	})
-	return err
-}
-
-func (r *organizationRepository) ReplaceOwner(ctx context.Context, tenant, organizationId, userId string) error {
-	span, ctx := opentracing.StartSpanFromContext(ctx, "OrganizationRepository.ReplaceOwner")
-	defer span.Finish()
-	tracing.SetNeo4jRepositorySpanTags(ctx, span, tenant)
-	span.LogFields(log.String("organizationId", organizationId), log.String("userId", userId))
-
-	query := `MATCH (t:Tenant {name:$tenant})<-[:ORGANIZATION_BELONGS_TO_TENANT]-(org:Organization {id:$organizationId})
-			OPTIONAL MATCH (:User)-[rel:OWNS]->(org)
-			DELETE rel
-			WITH org, t
-			MATCH (t)<-[:USER_BELONGS_TO_TENANT]-(u:User {id:$userId})
-			WHERE (u.internal=false OR u.internal is null) AND (u.bot=false OR u.bot is null)
-			MERGE (u)-[:OWNS]->(org)
-			SET org.updatedAt=$now, org.sourceOfTruth=$source`
-
-	session := utils.NewNeo4jWriteSession(ctx, *r.driver, utils.WithDatabaseName(r.database))
-	defer session.Close(ctx)
-
-	return utils.ExecuteWriteQuery(ctx, *r.driver, query, map[string]any{
-		"tenant":         tenant,
-		"organizationId": organizationId,
-		"userId":         userId,
-		"source":         constants.SourceOpenline,
-		"now":            utils.Now(),
-	})
-}
-
-func (r *organizationRepository) SetVisibility(ctx context.Context, tenant, organizationId string, hide bool) error {
-	span, ctx := opentracing.StartSpanFromContext(ctx, "OrganizationRepository.SetVisibility")
-	defer span.Finish()
-	tracing.SetNeo4jRepositorySpanTags(ctx, span, tenant)
-	span.LogFields(log.String("organizationId", organizationId), log.Bool("hide", hide))
-
-	query := fmt.Sprintf(`MATCH (t:Tenant {name:$tenant})<-[:ORGANIZATION_BELONGS_TO_TENANT]-(org:Organization:Organization_%s {id:$id})
-		 SET	org.hide = $hide,
-				org.updatedAt = $now`, tenant)
-
-	span.LogFields(log.String("query", query))
-
-	return r.executeQuery(ctx, query, map[string]any{
-		"id":     organizationId,
-		"tenant": tenant,
-		"hide":   hide,
-		"now":    utils.Now(),
-	})
-}
-
-func (r *organizationRepository) UpdateLastTouchpoint(ctx context.Context, tenant, organizationId string, touchpointAt time.Time, touchpointId, touchpointType string) error {
-	span, ctx := opentracing.StartSpanFromContext(ctx, "OrganizationRepository.SetVisibility")
-	defer span.Finish()
-	tracing.SetNeo4jRepositorySpanTags(ctx, span, tenant)
-	span.LogFields(log.String("organizationId", organizationId), log.String("touchpointId", touchpointId), log.Object("touchpointAt", touchpointAt))
-
-	query := `MATCH (:Tenant {name:$tenant})<-[:ORGANIZATION_BELONGS_TO_TENANT]-(org:Organization {id:$organizationId})
-		 SET org.lastTouchpointAt=$touchpointAt, org.lastTouchpointId=$touchpointId, org.lastTouchpointType=$touchpointType`
-	span.LogFields(log.String("query", query))
-
-	session := utils.NewNeo4jWriteSession(ctx, *r.driver, utils.WithDatabaseName(r.database))
-	defer session.Close(ctx)
-
-	return r.executeQuery(ctx, query, map[string]any{
-		"tenant":         tenant,
-		"organizationId": organizationId,
-		"touchpointAt":   touchpointAt,
-		"touchpointId":   touchpointId,
-		"touchpointType": touchpointType,
-	})
-}
-
-func (r *organizationRepository) SetCustomerOsIdIfMissing(ctx context.Context, tenant, organizationId, customerOsId string) error {
-	span, ctx := opentracing.StartSpanFromContext(ctx, "OrganizationRepository.SetCustomerOsIdIfMissing")
-	defer span.Finish()
-	tracing.SetNeo4jRepositorySpanTags(ctx, span, tenant)
-	span.LogFields(log.String("organizationId", organizationId), log.String("customerOsId", customerOsId))
-
-	query := `MATCH (:Tenant {name:$tenant})<-[:ORGANIZATION_BELONGS_TO_TENANT]-(org:Organization {id:$organizationId})
-		 SET org.customerOsId = CASE WHEN (org.customerOsId IS NULL OR org.customerOsId = '') AND $customerOsId <> '' THEN $customerOsId ELSE org.customerOsId END`
-	span.LogFields(log.String("query", query))
-
-	session := utils.NewNeo4jWriteSession(ctx, *r.driver, utils.WithDatabaseName(r.database))
-	defer session.Close(ctx)
-
-	return r.executeQuery(ctx, query, map[string]any{
-		"tenant":         tenant,
-		"organizationId": organizationId,
-		"customerOsId":   customerOsId,
-	})
-}
-
-func (r *organizationRepository) LinkWithParentOrganization(ctx context.Context, tenant, organizationId, parentOrganizationId, subOrganizationType string) error {
-	span, ctx := opentracing.StartSpanFromContext(ctx, "OrganizationRepository.LinkWithParentOrganization")
-	defer span.Finish()
-	tracing.SetNeo4jRepositorySpanTags(ctx, span, tenant)
-	span.LogFields(log.String("organizationId", organizationId), log.String("parentOrganizationId", parentOrganizationId), log.String("subOrganizationType", subOrganizationType))
-
-	query := `MATCH (t:Tenant {name:$tenant})<-[:ORGANIZATION_BELONGS_TO_TENANT]-(parent:Organization {id:$parentOrganizationId}),
-		 			(t)<-[:ORGANIZATION_BELONGS_TO_TENANT]-(sub:Organization {id:$subOrganizationId}) 
-		 	MERGE (sub)-[rel:SUBSIDIARY_OF]->(parent) 
-		 		ON CREATE SET rel.type=$type 
-		 		ON MATCH SET rel.type=$type`
-	span.LogFields(log.String("query", query))
-
-	return r.executeQuery(ctx, query, map[string]any{
-		"tenant":               tenant,
-		"subOrganizationId":    organizationId,
-		"parentOrganizationId": parentOrganizationId,
-		"type":                 subOrganizationType,
-	})
-}
-
-func (r *organizationRepository) UnlinkParentOrganization(ctx context.Context, tenant, organizationId, parentOrganizationId string) error {
-	span, ctx := opentracing.StartSpanFromContext(ctx, "OrganizationRepository.UnlinkParentOrganization")
-	defer span.Finish()
-	tracing.SetNeo4jRepositorySpanTags(ctx, span, tenant)
-	span.LogFields(log.String("organizationId", organizationId), log.String("parentOrganizationId", parentOrganizationId))
-
-	query := `MATCH (t:Tenant {name:$tenant})<-[:ORGANIZATION_BELONGS_TO_TENANT]-(parent:Organization {id:$parentOrganizationId})<-[rel:SUBSIDIARY_OF]-(sub:Organization {id:$subOrganizationId})-[:ORGANIZATION_BELONGS_TO_TENANT]->(t)
-		 		DELETE rel`
-	span.LogFields(log.String("query", query))
-
-	return r.executeQuery(ctx, query, map[string]any{
-		"tenant":               tenant,
-		"subOrganizationId":    organizationId,
-		"parentOrganizationId": parentOrganizationId,
-	})
-}
-
-func (r *organizationRepository) UpdateArr(ctx context.Context, tenant, organizationId string) error {
-	span, ctx := opentracing.StartSpanFromContext(ctx, "OrganizationRepository.UpdateArr")
-	defer span.Finish()
-	tracing.SetNeo4jRepositorySpanTags(ctx, span, tenant)
-	span.LogFields(log.String("organizationId", organizationId))
-
-	cypher := `MATCH (t:Tenant {name: $tenant})<-[:ORGANIZATION_BELONGS_TO_TENANT]-(org:Organization {id: $organizationId})
-				OPTIONAL MATCH (org)-[:HAS_CONTRACT]->(c:Contract)
-				OPTIONAL MATCH (c)-[:ACTIVE_RENEWAL]->(op:Opportunity)
-				WITH org, COALESCE(sum(op.amount), 0) as arr, COALESCE(sum(op.maxAmount), 0) as maxArr
-				SET org.renewalForecastArr = arr, org.renewalForecastMaxArr = maxArr, org.updatedAt = $now`
-	params := map[string]any{
-		"tenant":         tenant,
-		"organizationId": organizationId,
-		"now":            utils.Now(),
-	}
-	span.LogFields(log.String("query", cypher))
-	tracing.LogObjectAsJson(span, "params", params)
-
-	return r.executeQuery(ctx, cypher, params)
-}
-
-func (r *organizationRepository) UpdateRenewalSummary(ctx context.Context, tenant, organizationId string, likelihood *string, likelihoodOrder *int64, nextRenewalDate *time.Time) error {
-	span, ctx := opentracing.StartSpanFromContext(ctx, "OrganizationRepository.UpdateRenewalSummary")
-	defer span.Finish()
-	tracing.SetNeo4jRepositorySpanTags(ctx, span, tenant)
-	span.LogFields(log.String("organizationId", organizationId), log.Object("likelihood", likelihood), log.Object("likelihoodOrder", likelihoodOrder), log.Object("nextRenewalDate", nextRenewalDate))
-
-	cypher := `MATCH (t:Tenant {name:$tenant})<-[:ORGANIZATION_BELONGS_TO_TENANT]-(org:Organization {id:$organizationId})
-				SET org.derivedRenewalLikelihood = $derivedRenewalLikelihood,
-					org.derivedRenewalLikelihoodOrder = $derivedRenewalLikelihoodOrder,
-					org.derivedNextRenewalAt = $derivedNextRenewalAt,
-					org.updatedAt = $now`
-	params := map[string]any{
-		"tenant":                        tenant,
-		"organizationId":                organizationId,
-		"derivedRenewalLikelihood":      likelihood,
-		"derivedRenewalLikelihoodOrder": likelihoodOrder,
-		"derivedNextRenewalAt":          utils.TimePtrFirstNonNilNillableAsAny(nextRenewalDate),
-		"now":                           utils.Now(),
-	}
-	span.LogFields(log.String("query", cypher), log.Object("params", params))
-
-	return r.executeQuery(ctx, cypher, params)
-}
-
-func (r *organizationRepository) WebScrapeRequested(ctx context.Context, tenant, organizationId, url string, attempt int64, requestedAt time.Time) error {
-	span, ctx := opentracing.StartSpanFromContext(ctx, "OrganizationRepository.WebScrapeRequested")
-	defer span.Finish()
-	tracing.SetNeo4jRepositorySpanTags(ctx, span, tenant)
-	span.LogFields(log.String("organizationId", organizationId), log.String("url", url))
-
-	cypher := `MATCH (t:Tenant {name:$tenant})<-[:ORGANIZATION_BELONGS_TO_TENANT]-(org:Organization {id:$organizationId})
-		 	SET org.webScrapeLastRequestedAt=$requestedAt, 
-				org.webScrapeLastRequestedUrl=$url, 
-				org.webScrapeAttempts=$attempt`
-	params := map[string]any{
-		"tenant":         tenant,
-		"organizationId": organizationId,
-		"url":            url,
-		"attempt":        attempt,
-		"requestedAt":    requestedAt,
-	}
-	span.LogFields(log.String("cypher", cypher))
-	tracing.LogObjectAsJson(span, "params", params)
-
-	return r.executeQuery(ctx, cypher, params)
-}
-
-func (r *organizationRepository) UpdateOnboardingStatus(ctx context.Context, tenant, organizationId, status, comments string, statusOrder *int64, updatedAt time.Time) error {
-	span, ctx := opentracing.StartSpanFromContext(ctx, "OrganizationRepository.UpdateOnboardingStatus")
-	defer span.Finish()
-	tracing.SetNeo4jRepositorySpanTags(ctx, span, tenant)
-	span.SetTag(tracing.SpanTagEntityId, organizationId)
-
-	cypher := `MATCH (t:Tenant {name:$tenant})<-[:ORGANIZATION_BELONGS_TO_TENANT]-(org:Organization {id:$organizationId})
-				SET org.onboardingUpdatedAt = CASE WHEN org.onboardingStatus IS NULL OR org.onboardingStatus <> $status THEN $updatedAt ELSE org.onboardingUpdatedAt END,
-					org.onboardingStatus=$status,
-					org.onboardingStatusOrder=$statusOrder,
-					org.onboardingComments=$comments,
-					org.onboardingUpdatedAt=$updatedAt,
-					org.updatedAt=$now`
-	params := map[string]any{
-		"tenant":         tenant,
-		"organizationId": organizationId,
-		"status":         status,
-		"statusOrder":    statusOrder,
-		"comments":       comments,
-		"updatedAt":      updatedAt,
-		"now":            utils.Now(),
-	}
-	span.LogFields(log.String("cypher", cypher))
-	tracing.LogObjectAsJson(span, "params", params)
-
-	err := r.executeQuery(ctx, cypher, params)
-	if err != nil {
-		tracing.TraceErr(span, err)
-	}
-	return err
-}
-
-// Common database interaction method
-func (r *organizationRepository) executeQuery(ctx context.Context, query string, params map[string]any) error {
-	return utils.ExecuteWriteQuery(ctx, *r.driver, query, params)
-}
