+++ packages/server/events-processing-platform/repository/repositories.go
@@ -19,8 +19,6 @@ type Repositories struct {
 	Neo4jRepositories       *neo4jrepository.Repositories
 	CommonRepositories      *cmn_repository.Repositories
 	CustomerOsIdsRepository repository.CustomerOsIdsRepository
-
-	OrganizationRepository OrganizationRepository
 }
 
 func InitRepos(driver *neo4j.DriverWithContext, neo4jDatabase string, gormDb *gorm.DB) *Repositories {
@@ -31,7 +29,6 @@ func InitRepos(driver *neo4j.DriverWithContext, neo4jDatabase string, gormDb *go
 		Neo4jRepositories:       neo4jrepository.InitNeo4jRepositories(driver, neo4jDatabase),
 		CommonRepositories:      cmn_repository.InitRepositories(gormDb, driver),
 		CustomerOsIdsRepository: repository.NewCustomerOsIdsRepository(gormDb),
-		OrganizationRepository:  NewOrganizationRepository(driver, neo4jDatabase),
 	}
 
 	err := gormDb.AutoMigrate(&entity.CustomerOsIds{})
